+++ core/src/main/java/react4j/core/ReactRefCallback.java
@@ -1,9 +1,0 @@
-package react4j.core;
-
-import jsinterop.annotations.JsFunction;
-
-@JsFunction
-public interface ReactRefCallback
-{
-  void passRef( Object refElement );
-}