+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 * 💥 **\[dom\]** Change the event handlers on `LabelProps` to use camel case rather than pascal case. i.e. Rename the
   fluent event handlers from `OnX` to `onX`
 * 💥 **\[core\]** Move `react.dom.RefConsumer` to `react.core.RefConsumer`.
+* 💥 **\[core\]** Remove support for string refs as they are deprecated in React js and slated for removal.
+* 💥 **\[core\]** Remove `ReactRefCallback` and replace it with `RefConsumer`.
 
 ### [v0.12](https://github.com/realityforge/react4j/tree/v0.12) (2017-11-09)
 [Full Changelog](https://github.com/realityforge/react4j/compare/v0.11...v0.12)